requirements.txt
future==0.18.2
requests==2.27.1
Pillow==9.0.1
selenium==4.1.2
rich==11.2.0
selenium-wire==4.6.2